fix: The applet list is not translated.
